update : service 구현 방식을 interface + impl 구조로 변경
